ide and harms inconsistency

+++ api/ArkUI/component/ohos.arkui.component.gauge.cj.d
@@ -19,52 +19,52 @@ public class Gauge <: CommonMethodComponent<Gauge> & GaugeAttribute {
 /**
  * The constructor of Gauge.
  *
- * @param { ?Float64 } value - Set current data value.
- * @param { ?Float64 } [min] - Set current segment minimum value.
- * @param { ?Float64 } [max] - Set current segment maximum value.
+ * @param { ?Float32 } value - Set current data value.
+ * @param { ?Float32 } [min] - Set current segment minimum value.
+ * @param { ?Float32 } [max] - Set current segment maximum value.
  * @param { () -> Unit } [child] - Set child component of Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    public init(value!: ?Float64, min!: ?Float64 = None, max!: ?Float64 = None, child!: () -> Unit = {=>})
+    public init(value!: ?Float32, min!: ?Float32 = None, max!: ?Float32 = None, child!: () -> Unit = {=>})
 
 /**
  * Sets the value for the current profile.
  *
- * @param { ?Float64 } value - Set current data value.
+ * @param { ?Float32 } value - Set current data value.
  * @returns { This } returns the instance of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    public func value(value: ?Float64): This
+    public func value(value: ?Float32): This
 
 /**
  * Set the start angle. Clock 0 is 0 degrees and clockwise is positive.
  *
- * @param { ?Float64 } angle - The start angle value, measured in degrees.
+ * @param { ?Float32 } angle - The start angle value, measured in degrees.
  * @returns { This } returns the instance of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    public func startAngle(angle: ?Float64): This
+    public func startAngle(angle: ?Float32): This
 
 /**
  * Sets the end angle position. Clock 0 is 0 degrees and clockwise is positive.
  *
- * @param { ?Float64 } angle - The end angle value, measured in degrees.
+ * @param { ?Float32 } angle - The end angle value, measured in degrees.
  * @returns { This } returns the instance of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    public func endAngle(angle: ?Float64): This
+    public func endAngle(angle: ?Float32): This
 
 /**
  * Set the color of the chart. You can set the solid color and segmented gradient color.
@@ -129,16 +129,16 @@ public class Gauge <: CommonMethodComponent<Gauge> & GaugeAttribute {
 /**
  * Sets track shadow of the ring chart.
  *
- * @param { ?Float64 } [radius] - The shadow radius.
- * @param { ?Float64 } [offsetX] - The shadow horizontal offset.
- * @param { ?Float64 } [offsetY] - The shadow vertical offset.
+ * @param { ?Float32 } [radius] - The shadow radius.
+ * @param { ?Float32 } [offsetX] - The shadow horizontal offset.
+ * @param { ?Float32 } [offsetY] - The shadow vertical offset.
  * @returns { This } returns the instance of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    public func trackShadow(radius!: ?Float64 = None, offsetX!: ?Float64 = None, offsetY!: ?Float64 = None): This
+    public func trackShadow(radius!: ?Float32 = None, offsetX!: ?Float32 = None, offsetY!: ?Float32 = None): This
 
 /**
  * Sets indicator options of the ring chart.
@@ -189,26 +189,26 @@ sealed interface GaugeAttribute <: CommonMethod<GaugeAttribute> {
 /**
  * Set the start angle. Clock 0 is 0 degrees and clockwise is positive.
  *
- * @param { ?Float64 } angle - The start angle value.
+ * @param { ?Float32 } angle - The start angle value.
  * @returns { GaugeAttribute } The attribute of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    func startAngle(angle: ?Float64): GaugeAttribute
+    func startAngle(angle: ?Float32): GaugeAttribute
 
 /**
  * Sets the end angle position. Clock 0 is 0 degrees and clockwise is positive.
  *
- * @param { ?Float64 } angle - The end angle value.
+ * @param { ?Float32 } angle - The end angle value.
  * @returns { GaugeAttribute } The attribute of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    func endAngle(angle: ?Float64): GaugeAttribute
+    func endAngle(angle: ?Float32): GaugeAttribute
 
 /**
  * Set the color of the chart. You can set the solid color and segmented gradient color.
@@ -273,16 +273,16 @@ sealed interface GaugeAttribute <: CommonMethod<GaugeAttribute> {
 /**
  * Sets track shadow of the ring chart.
  *
- * @param { ?Float64 } radius - The shadow radius.
- * @param { ?Float64 } offsetX - The shadow horizontal offset.
- * @param { ?Float64 } offsetY - The shadow vertical offset.
+ * @param { ?Float32 } radius - The shadow radius.
+ * @param { ?Float32 } offsetX - The shadow horizontal offset.
+ * @param { ?Float32 } offsetY - The shadow vertical offset.
  * @returns { GaugeAttribute } The attribute of the Gauge.
  */
 @!APILevel[
     since: "22",
     syscap: "SystemCapability.ArkUI.ArkUI.Full"
 ]
-    func trackShadow(radius!: ?Float64, offsetX!: ?Float64, offsetY!: ?Float64): GaugeAttribute
+    func trackShadow(radius!: ?Float32, offsetX!: ?Float32, offsetY!: ?Float32): GaugeAttribute
 
 /**
  * Sets indicator options of the ring chart.
